lib: Remove implicit global initialization and add note about thread safety
Without adding explicit global initialization routines to libresdet and assuming
the presence of threaded FFTW, we cannot guarantee thread safety when
FFTW and ImageMagick support is enabled.

Since these are optional dependencies, the responsibility of global prep for
threaded execution of these is instead deferred to the application itself, which
is not ideal but at least allows for concurrency in those configurations.