Use proper u16 macro for scanf
The PRIu16 macro only guarantees that the resulting conversion
specifier will be suitable for printing, not scanning.
SCNu16 provides the appropriate specifier for the latter.

Some libcs define PRIu16 to just "%u" rather than defining both as
"%hu", which is makes sense but results in undefined behavior when
used with scanf.