Prefer __builtin_signbit when available.
This is to work around a quirk specific to macOS that warrants
some explanation:
On macOS several math builtins, including those for signbit, are omitted
in favor of their library equivalent when __FINITE_MATH_ONLY__ is set
(e.g. via -ffast-math). This is done to ensure correct behavior in the
presence of NaNs and Inf, but it means that counter-intuitively
-ffast-math may actually result in slower generated code as these
functions won't be inlined.

It's not currently possible for resdet to produce non-finite numbers
as transform coefficients except in the event of memory corruption,
but if it were, for IEEE 754 at least these have the same semantics for
signbit's purposes as finite numbers.
The former behavior is still available if needed by configuring with
the --no-builtin-signbit flag.

With -Ofast this change results in between 1.3x and 2x speedup on macOS
depending on the detection method, with the default method "sign" seeing
the largest gain.

glibc doesn't make this distinction and already used the faster builtin.