libresdet: Switch to float for input image buffers
This breaking change is motivated by the fact that resdet performs
substantially better (up to +15% confidence with the default method)
when detecting in the colorspace the input was originally resized in.
This allows clients to provide higher precision inputs that have
undergone colorspace conversions.

This also adds portable floatmap support for resdet's built-in image
readers to enable reading higher precision images using the command-
line tool.
This reader supports a series of PFM images for multiple frame input
as well in the form that FFmpeg and ImageMagick both use for writing
PFM sequences.