Consolidate overflow checks for input image dimensions
Adds an internal macro to check if the given dimensions would either
exceed the PIXEL_MAX limit or would overflow on multiplication, which
replaces all previous overflow checks.

This fixes an issue introduced in 36636cc005680889cb34b0321ca1752a5ee18044
where overflow would not be caught if the product of an input image's
dimensions was less than SIZE_MAX but greater than SIZE_MAX divided by
the size of a float.