lib: allocate exact length for output resolution arrays
Previously these were allocated ahead of time to a maximum length based
on the input dimensions, but this is likely to be far larger than
the actual needed length any time threshold is > 0.

Instead calculate the real result set size and allocate based on that.

There is no change for callers other than slightly reduced memory use.