Use symbolic defines for precision levels in precision.h
Affects the value given to --coeff-precision and --inter-precision when
configuring, which is now F, D, or L for float, double, or long
double respectively.

The previous numeric values were based on x87 byte widths, but these are
no longer representative as x87 floating point processing is not in wide
use anymore.